+++ api/osb/check_visibility_plugin.go
@@ -1,4 +1,4 @@
-package plugins
+package osb
 
 import (
 	"encoding/json"
